27/10/2019--nv

+++ app/Http/Controllers/DonHangController.php
@@ -55,8 +55,15 @@ class DonHangController extends Controller
         }
     }
 
-    public function delete($id){
-        $donhang = DonHang::findOrfail($id);
-        $donhang->delete();
+    public function delete($id)
+    {
+        if (\Session::get('quyen') == 1) {
+            $donhang = DonHang::findOrfail($id);
+            $donhang->delete();
+        }
+        else {
+            return response(["error" => 'Bạn không có quyền xóa']);
+        }
     }
+
 }
