+++ app/Http/Controllers/GiongController.php
@@ -39,8 +39,14 @@ class GiongController extends Controller
         }
     }
 
-    public function delete($id){
-        $giong = Giong::findOrfail($id);
-        $giong->delete();
+    public function delete($id)
+    {
+        if (\Session::get('quyen') == 1) {
+            $giong = Giong::findOrfail($id);
+            $giong->delete();
+        }
+        else {
+            return response(["error" => 'Bạn không có quyền xóa']);
+        }
     }
 }
